Update for SL4
This is an example how this could work in SL4.

`'-jar'` if you prefix a default with e.g. `-` it will be recognized as cli argument and injected at `${args}`. Likewise for `-configfile`. Esp. for the configfile the user could point to `${folder}/.myconfig.xml` or `~/.myotherconfig.xml`. `$folder` will expand to the project folder.